Correccion de repos, añadir service, impl y controlador del cliente

+++ Servidor/src/main/java/com/example/demo/Repository/IngredienteProductoRepository.java
@@ -2,10 +2,9 @@ package com.example.demo.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 import com.example.demo.Entity.IngredienteProductoEntity;
 
 @Repository
-public interface IngredienteProductoRepository extends JpaRepository<IngredienteProductoEntity, Integer>{
+public interface IngredienteProductoRepository extends JpaRepository<IngredienteProductoEntity, Long>{
 
 }
